Hardcoded a session with password

+++ index.php
@@ -1,4 +1,24 @@
 <?php
+session_start();
+
+if (
+    isset($_POST['login'])
+    && !empty($_POST['username'])
+    && !empty($_POST['password'])
+) {
+    if (
+        $_POST['username'] == 'ern' &&
+        $_POST['password'] == 'zefyras'
+    ) {
+        $_SESSION['logged_in'] = true;
+        $_SESSION['timeout'] = time();
+        $_SESSION['username'] = 'ern';
+    }
+}
+
+?>
+
+<?php
 # Defining main globals for path managing
 $root_dir = getcwd();
 $sep = DIRECTORY_SEPARATOR;
@@ -6,6 +26,7 @@ $tmp = explode($sep, $root_dir);
 #Making base directory one level up
 $dir = array_splice($tmp, count($tmp) - 1);
 $base_dir = implode($sep, $tmp);
+
 
 ?>
 <!DOCTYPE html>
@@ -33,25 +54,44 @@ $base_dir = implode($sep, $tmp);
             <div class="clouds"></div>
             <h3 class="error" id="new_dir_error"></h3>
             <?php
+            if (isset($_GET['action']) and $_GET['action'] == 'logout') {
+                session_start();
+                unset($_SESSION['username']);
+                unset($_SESSION['password']);
+                unset($_SESSION['logged_in']);
+                echo '<h3 class="utility__message">You have logged out</h3>';
+                echo '<div class="utility">
+                <a class="utility__link" href="index.html">Back to main</a>
+                </div>"
+                ';
+            }
+            ?>
+
+
+
+
+            <?php
 
             if ($_SERVER["REQUEST_METHOD"] === "POST") {
 
-                $rel = $_POST['file'];
+
+                if ($_SESSION['logged_in'] === true) {
+                    $rel = $_POST['file'];
+                    // if we are atleast 1 level deep updating base directory
+                    $base_dir = $rel ? $base_dir . $rel : $base_dir;
+
+                    # Confirm delete
+                    if ($_POST['confirmDelete']) {
+                        $path_to_item = $_POST['confirmDelete'];
+                        print_r($path_to_item);
+                        unlink($path_to_item);
+                    }
+
+                    echo "<ul>";
 
 
-                // if we are atleast 1 level deep updating base directory
-                $base_dir = $rel ? $base_dir . $rel : $base_dir;
 
-                # Confirm delete
-                if ($_POST['confirmDelete']) {
-                    $path_to_item = $_POST['confirmDelete'];
-                    print_r($path_to_item);
-                    unlink($path_to_item);
-                }
-
-                echo "<ul>";
-
-                echo '<li>
+                    echo '<li>
                 <form action="" method="POST">
                 <input type="hidden" name="file" value="' . $rel . '">
                 <input type="hidden" name="uploadModal" value="openUploadModal">
@@ -63,8 +103,8 @@ $base_dir = implode($sep, $tmp);
 
 
 
-                # New directory
-                echo '<li>
+                    # New directory
+                    echo '<li>
                 <form action="" method="POST" id="new_dir_form">
                 <input type="hidden" name="file" value="' . $rel . '">
                <div class="input-container new_dir">
@@ -86,22 +126,22 @@ $base_dir = implode($sep, $tmp);
                 </div></li>
                 ';
 
-                if ($_POST['new_dir']) {
-                    $new_dir = $_POST['new_dir'];
-                    mkdir($base_dir . $sep . $new_dir);
-                }
-                if ($rel) {
-                    # Generating a back button with a back path if atleast 1 level deep
-                    # Back path is basically the path - last elementy after /
-                    $back_path = explode($sep, $rel);
-                    $back_path = array_filter($back_path);
-                    array_pop($back_path);
-                    $back_path = implode($sep, $back_path);
-                    if ($back_path) {
-                        $back_path = $sep . $back_path;
+                    if ($_POST['new_dir']) {
+                        $new_dir = $_POST['new_dir'];
+                        mkdir($base_dir . $sep . $new_dir);
                     }
-                    echo "<br>";
-                    echo '
+                    if ($rel) {
+                        # Generating a back button with a back path if atleast 1 level deep
+                        # Back path is basically the path - last elementy after /
+                        $back_path = explode($sep, $rel);
+                        $back_path = array_filter($back_path);
+                        array_pop($back_path);
+                        $back_path = implode($sep, $back_path);
+                        if ($back_path) {
+                            $back_path = $sep . $back_path;
+                        }
+                        echo "<br>";
+                        echo '
                     <li> <form action="index.php" method="POST">
             <input type="hidden" name="file" value="' . $back_path . '">
             <button class="back" type="submit">
@@ -109,41 +149,41 @@ $base_dir = implode($sep, $tmp);
             Go back </button>
             </form></li>
             ';
-                }
-                // Getting content from new directory
-                // Printing all directories
-                $content = scandir($base_dir);
-                for ($i = 2; $i < count($content); $i++) {
-                    # Define new current directory
-                    $curr_dir = $base_dir . $sep . $content[$i];
-                    if (is_dir($curr_dir)) {
-                        #If it's a directory making a form with new path..
-                        echo '
+                    }
+                    // Getting content from new directory
+                    // Printing all directories
+                    $content = scandir($base_dir);
+                    for ($i = 2; $i < count($content); $i++) {
+                        # Define new current directory
+                        $curr_dir = $base_dir . $sep . $content[$i];
+                        if (is_dir($curr_dir)) {
+                            #If it's a directory making a form with new path..
+                            echo '
                         <li data-search="' . $content[$i] . '"><form action="index.php"  method="POST" >
             <input type="hidden" name="file" value="' .
-                            $rel . $sep . $content[$i] .
-                            '"/>
+                                $rel . $sep . $content[$i] .
+                                '"/>
             <img class="browser_img" alt="folder" src="assets/img/folder.png">
             <button type="submit" class="folder">
             ' . $content[$i] . ' </button>
             </form></li>
             ';
+                        }
                     }
-                }
-                #printing all files
-                for ($i = 2; $i < count($content); $i++) {
+                    #printing all files
+                    for ($i = 2; $i < count($content); $i++) {
 
-                    if (is_file($base_dir . $sep . $content[$i])) {
-                        $curr_dir = $base_dir . $sep . $content[$i];
-                        # Getting file extension to show correct image
-                        $split = explode('.', $content[$i]);
-                        $ext = $split[count($split) - 1];
-                        # Getting the relative path and putting it on href
-                        $path = $rel ? $rel . $sep . $content[$i] : $content[$i];
-                        # Need to trim the first slash so it refers to root directory
-                        $path = ltrim($path, $sep);
-                        # File with all its options (Delete, download..)..
-                        echo '<li data-search="' . $content[$i] . '"><div class="file">
+                        if (is_file($base_dir . $sep . $content[$i])) {
+                            $curr_dir = $base_dir . $sep . $content[$i];
+                            # Getting file extension to show correct image
+                            $split = explode('.', $content[$i]);
+                            $ext = $split[count($split) - 1];
+                            # Getting the relative path and putting it on href
+                            $path = $rel ? $rel . $sep . $content[$i] : $content[$i];
+                            # Need to trim the first slash so it refers to root directory
+                            $path = ltrim($path, $sep);
+                            # File with all its options (Delete, download..)..
+                            echo '<li data-search="' . $content[$i] . '"><div class="file">
                             <img class="file_img" alt="file" src="assets/img/' . $ext . '.png">
                             <a class="openFile" href="/' . $path  . ' "> ' . $content[$i] . '  </a>
                             <form action="" method="POST">
@@ -157,17 +197,17 @@ $base_dir = implode($sep, $tmp);
                             </button>
                             </form>
                 </div></li>';
+                        }
                     }
-                }
 
 
-                #Delete
-                if ($_POST['delete']) {
-                    $delete_path = $_POST['delete'];
-                    $temp = explode($sep, $delete_path);
-                    $fileName = $temp[count($temp) - 1];
+                    #Delete
+                    if ($_POST['delete']) {
+                        $delete_path = $_POST['delete'];
+                        $temp = explode($sep, $delete_path);
+                        $fileName = $temp[count($temp) - 1];
 
-                    echo '<div class="modal_overlay modal_overlay--active" id="confirmDeleteModal">
+                        echo '<div class="modal_overlay modal_overlay--active" id="confirmDeleteModal">
                     <form action="" method="POST" id="confirmDelete" class="modal">
                     <h3>Are you sure you want to delete <span>' . $fileName . '</span> ? </h3>
                     <input type="hidden" name="file" value="' . $rel . '">
@@ -179,10 +219,10 @@ $base_dir = implode($sep, $tmp);
                     </form>
                 </div>
                     ';
-                }
-                #Upload
-                if ($_POST['uploadModal']) {
-                    echo ' <div class="modal_overlay modal_overlay--active" id="uploadModal">
+                    }
+                    #Upload
+                    if ($_POST['uploadModal']) {
+                        echo ' <div class="modal_overlay modal_overlay--active" id="uploadModal">
 
                     <form action="upload.php" method="POST" id="uploadForm" class="modal" enctype="multipart/form-data">
                     <button id="closeUploadModal">Close</button>
@@ -195,16 +235,21 @@ $base_dir = implode($sep, $tmp);
                     Submit upload</button>
                     </form>
                     ';
+                    }
+                    echo "</ul>";
+                    # User UI corner
+
+                    echo '<div class="userUI">
+                    <h4 class="userUI__name">Welcome, ' . $_SESSION['username'] . '!</h4>
+                    <a class="userUI__link" href="index.php?action=logout">Logout</a>
+                    </div>';
+                } else {
+                    echo '<h3 class="utility__message utility__message--red"> Sorry, something went wrong.</h3>
+                    <div class="utility">
+                    <a class="utility__link" href="index.html">Back to main</a>
+                    </div>
+                    ';
                 }
-
-
-                echo "</ul";
-            } else {
-                echo '<h3 class="utility__message error"> Sorry, something went wrong.</h3>
-                <form action="index.html" method="POST" class="utility">
-                <button class="squareBtn" type="submit">Back to main</button>
-                </form>
-                ';
             }
             ?>
         </div>
